WIP fixing more broken things

+++ api/src/endpoints/websocket.py
@@ -1,6 +1,7 @@
 from starlette.endpoints import WebSocketEndpoint
 from starlette.websockets import WebSocket
 from pydantic import ValidationError
+from lagom import injectable
 from ..services import WebSocketService
 from ..validator import SubscriptionActionValidator
 from ..utils.constants import SUBSCRIBE, UNSUBSCRIBE
@@ -9,22 +10,29 @@ from typing import Any
 
 class UpdateStreamEndpoint(WebSocketEndpoint):
     async def on_connect(
-        self, socket: WebSocket, websocket_service: WebSocketService
+        self, websocket: WebSocket, websocket_service: WebSocketService = injectable
     ) -> None:
-        connection = websocket_service.create(socket)
+        connection = websocket_service.create(websocket)
 
         await connection.socket.send_json({"message": "Connected!!!"})
 
     async def on_receive(
-        self, socket: WebSocket, data: Any, websocekt_service: WebSocketService
+        self,
+        websocket: WebSocket,
+        data: Any,
+        websocekt_service: WebSocketService = injectable,
     ) -> None:
-        connection = websocekt_service.get_by_id(socket.state["id"])
+        connection = websocekt_service.get_by_id(websocket.state.id)
+
+        if connection is None:
+            return
 
         try:
             validator_data = SubscriptionActionValidator.model_validate(data)
+            topic_str = str(validator_data.topic)
 
             if validator_data.action.value == SUBSCRIBE:
-                if connection.topics.has(validator_data.topic):
+                if connection.topics.has(topic_str):
                     await connection.socket.send_json(
                         {
                             "event": "SUBSCRIBE:ALREADY_EXISTS",
@@ -35,9 +43,10 @@ class UpdateStreamEndpoint(WebSocketEndpoint):
                     )
                     return
 
-                connection = websocekt_service.subscribe(
-                    connection.id, validator_data.topic
-                )
+                connection = websocekt_service.subscribe(connection.id, topic_str)
+
+                if connection is None:
+                    return
 
                 await connection.socket.send_json(
                     {
@@ -49,9 +58,10 @@ class UpdateStreamEndpoint(WebSocketEndpoint):
                 )
 
             if validator_data.action.value == UNSUBSCRIBE:
-                connection = websocekt_service.unsubscribe(
-                    connection.id, validator_data.topic
-                )
+                connection = websocekt_service.unsubscribe(connection.id, topic_str)
+
+                if connection is None:
+                    return
 
                 await connection.socket.send_json(
                     {
@@ -63,6 +73,9 @@ class UpdateStreamEndpoint(WebSocketEndpoint):
                 )
 
         except ValidationError as err:
+            if connection is None:
+                return
+
             await connection.socket.send_json(
                 {
                     "event": "ERRORS:VALIDATION_ERROR",
@@ -74,9 +87,15 @@ class UpdateStreamEndpoint(WebSocketEndpoint):
             )
 
     async def on_disconnect(
-        self, socket: WebSocket, close_code: int, websocket_service: WebSocketService
+        self,
+        websocket: WebSocket,
+        close_code: int,
+        websocket_service: WebSocketService = injectable,
     ) -> None:
-        connection = websocket_service.get_by_id(socket.state["id"])
+        connection = websocket_service.get_by_id(websocket.state.id)
+
+        if connection is None:
+            return
 
         await connection.socket.send_json({"message": "Closing connection..."})
         websocket_service.delete(connection.id)
